Moved tests to lke-summary-page.spec.ts
